better init command

+++ main.go
@@ -34,31 +34,12 @@ func main() {
 		log.Fatal(err)
 		return
 	}
-	_, e := os.ReadFile(fmt.Sprintf("%s/cpkgs.json", dir))
-	if os.IsNotExist(e) {
-		var JSON JSON
-		JSON.Include.C = []string{}
-		JSON.Include.H = []string{}
-		j, _ := json.Marshal(JSON)
-		os.WriteFile("cpkgs.json", j, 0777)
-	}
 	var JSON JSON
 	j, _ := os.ReadFile(fmt.Sprintf("%s/cpkgs.json", dir))
-	err = json.Unmarshal(j, &JSON)
-	if err != nil {
-		JSON.Include.C = []string{}
-		JSON.Include.H = []string{}
-		j, _ := json.Marshal(JSON)
-		os.WriteFile("cpkgs.json", j, 0777)
-	}
+	json.Unmarshal(j, &JSON)
 	switch flag.Arg(0) {
 	case "run":
 		{
-			_, err := os.Getwd()
-			if err != nil {
-				log.Fatal(err)
-				return
-			}
 			f := filepath.Clean(flag.Arg(1))
 			file, err := filepath.Abs(f)
 			if err != nil {
@@ -70,7 +51,15 @@ func main() {
 				JSON.FileName += ".exe"
 			}
 			fname := JSON.FileName
-			cmd := fmt.Sprintf("cd %s && %s -o %s %s", path, JSON.Compiler, fname, strings.Join(flag.Args()[1:], " "))
+			_, err = os.Stat("cpkgs/bin")
+			if os.IsNotExist(err) {
+				err = os.Mkdir("cpkgs/bin", 0777)
+				if err != nil {
+					log.Fatal(err)
+					return
+				}
+			}
+			cmd := fmt.Sprintf("cd %s && %s -o cpkgs/bin/%s %s", path, JSON.Compiler, fname, strings.Join(flag.Args()[1:], " "))
 			files, err := os.ReadDir(fmt.Sprintf("%s/cpkgs", path))
 			if err != nil {
 				log.Fatal(err)
@@ -81,7 +70,7 @@ func main() {
 					cmd += fmt.Sprintf(" cpkgs/%s", files[i].Name())
 				}
 			}
-			cmd += fmt.Sprintf(" && %s", JSON.FileName)
+			cmd += fmt.Sprintf(" && cd cpkgs/bin && %s", JSON.FileName)
 			cmdExec := exec.Command("sh", "-c", cmd)
 			if runtime.GOOS == "windows" {
 				cmdExec = exec.Command("cmd", "/C", cmd)
@@ -95,6 +84,13 @@ func main() {
 	case "add":
 		{
 			pkgs := flag.Args()[1:]
+			if len(pkgs) <= 0 {
+				fmt.Print("Provide packages to add: ")
+				scanner := bufio.NewScanner(os.Stdin)
+				scanner.Scan()
+				p := scanner.Text()
+				pkgs = strings.Split(p, " ")
+			}
 			scanner := bufio.NewScanner(os.Stdin)
 			for i := 0; i < len(pkgs); i++ {
 				u, _ := url.Parse(pkgs[i])
@@ -106,12 +102,22 @@ func main() {
 					log.Fatal("Currently only github is supported!")
 					return
 				}
-				fmt.Printf("Specify headers file to add from '%s': ", strings.TrimPrefix(strings.ReplaceAll(u.Path, u.Host, ""), "/"))
+				fmt.Printf("Provide headers file to add from '%s': ", strings.TrimPrefix(strings.ReplaceAll(u.Path, u.Host, ""), "/"))
 				u.Host = "raw.githubusercontent.com"
 				urlString := strings.ReplaceAll(u.String(), "/github.com", "")
 				scanner.Scan()
 				h := scanner.Text()
 				if len(strings.TrimSpace(h)) <= 0 {
+					continue
+				}
+				found := false
+				for i := 0; i < len(JSON.Include.H) && !found; i++ {
+					url := strings.Split(JSON.Include.H[i], "/")
+					if h == url[len(url)-1] {
+						found = true
+					}
+				}
+				if found {
 					continue
 				}
 				headers := strings.Split(h, " ")
@@ -157,7 +163,11 @@ func main() {
 							return
 						}
 					}
-					os.WriteFile(fmt.Sprintf("cpkgs/%s", headers[i]), body, 0777)
+					err = os.WriteFile(fmt.Sprintf("cpkgs/%s", headers[i]), body, 0777)
+					if err != nil {
+						log.Fatal(err)
+						return
+					}
 					JSON.Include.H = append(JSON.Include.H, res.Request.URL.String())
 					code := strings.ReplaceAll(headers[i], ".h", ".c")
 					res, err = http.Get(fmt.Sprintf("%s/master/%s", urlString, strings.ReplaceAll(headers[i], ".h", ".c")))
@@ -173,42 +183,89 @@ func main() {
 						log.Fatal(err)
 						return
 					}
-					os.WriteFile(fmt.Sprintf("cpkgs/%s", code), body, 0777)
+					err = os.WriteFile(fmt.Sprintf("cpkgs/%s", code), body, 0777)
+					if err != nil {
+						log.Fatal(err)
+						return
+					}
 					JSON.Include.C = append(JSON.Include.C, res.Request.URL.String())
 					j, err := json.Marshal(JSON)
 					if err != nil {
 						log.Fatal(err)
 						return
 					}
-					os.WriteFile("cpkgs.json", j, 0777)
+					err = os.WriteFile("cpkgs.json", j, 0777)
+					if err != nil {
+						log.Fatal(err)
+						return
+					}
 				}
 			}
 			break
 		}
 	case "init":
 		{
-			os.WriteFile("cpkgs.json", j, 0777)
-			var compiler, filename string
-			fmt.Print("Specify the compiler to use: ")
-			fmt.Scanln(&compiler)
-			if len(strings.TrimSpace(compiler)) <= 0 {
-				fmt.Println("Using default...(gcc)")
-				compiler = "gcc"
+			dir, err := os.Getwd()
+			if err != nil {
+				log.Fatal(err)
+				return
 			}
-			JSON.Compiler = compiler
-			fmt.Print("Specify the output filename: ")
-			fmt.Scanln(&filename)
-			if len(strings.TrimSpace(filename)) <= 0 {
-				fmt.Println("Using default...(out)")
+			d := flag.Arg(1)
+			if d != "-d" && len(strings.TrimSpace(flag.Arg(1))) > 0 {
+				dir = flag.Arg(1)
+				d = flag.Arg(2)
+			}
+			_, e := os.Stat(dir)
+			if os.IsNotExist(e) {
+				os.Mkdir(dir, 0777)
+			}
+			_, e = os.ReadFile(fmt.Sprintf("%s/cpkgs.json", dir))
+			if os.IsNotExist(e) {
+				JSON.Include.C = []string{}
+				JSON.Include.H = []string{}
+				j, _ := json.Marshal(JSON)
+				err = os.WriteFile("cpkgs.json", j, 0777)
+				if err != nil {
+					log.Fatal(err)
+					return
+				}
+			}
+			err = os.WriteFile(fmt.Sprintf("%s/cpkgs.json", dir), j, 0777)
+			if err != nil {
+				log.Fatal(err)
+				return
+			}
+			var compiler, filename string
+			if d != "-d" {
+				fmt.Print("Provide the compiler to use: ")
+				fmt.Scanln(&compiler)
+				if len(strings.TrimSpace(compiler)) <= 0 {
+					fmt.Println("Using default...(gcc)")
+					compiler = "gcc"
+				}
+				fmt.Print("Provide the output filename: ")
+				fmt.Scanln(&filename)
+				if len(strings.TrimSpace(filename)) <= 0 {
+					fmt.Println("Using default...(out)")
+					filename = "out"
+				}
+			} else {
+				fmt.Println("Using defaults...(gcc, out)")
+				compiler = "gcc"
 				filename = "out"
 			}
+			JSON.Compiler = compiler
 			JSON.FileName = filename
 			j, err := json.Marshal(JSON)
 			if err != nil {
 				log.Fatal(err)
 				return
 			}
-			os.WriteFile("cpkgs.json", j, 0777)
+			err = os.WriteFile(fmt.Sprintf("%s/cpkgs.json", dir), j, 0777)
+			if err != nil {
+				log.Fatal(err)
+				return
+			}
 			break
 		}
 	case "install":
@@ -234,7 +291,11 @@ func main() {
 					return
 				}
 				filename := strings.Split(JSON.Include.H[i], "/")
-				os.WriteFile(fmt.Sprintf("%s/cpkgs/%s", dir, filename[len(filename)-1]), body, 0777)
+				err = os.WriteFile(fmt.Sprintf("%s/cpkgs/%s", dir, filename[len(filename)-1]), body, 0777)
+				if err != nil {
+					log.Fatal(err)
+					return
+				}
 				res, err = http.Get(JSON.Include.C[i])
 				if err != nil {
 					log.Fatal(err)
@@ -247,7 +308,11 @@ func main() {
 					return
 				}
 				filename = strings.Split(JSON.Include.C[i], "/")
-				os.WriteFile(fmt.Sprintf("%s/cpkgs/%s", dir, filename[len(filename)-1]), body, 0777)
+				err = os.WriteFile(fmt.Sprintf("%s/cpkgs/%s", dir, filename[len(filename)-1]), body, 0777)
+				if err != nil {
+					log.Fatal(err)
+					return
+				}
 			}
 			break
 		}
@@ -256,13 +321,14 @@ func main() {
 			fmt.Println("List of all commands:")
 			fmt.Print("\n")
 			fmt.Println("---------------------------------------------------------------------------")
-			fmt.Println("|'cpkgs init' - initialize a new project using cpkgs                      |")
+			fmt.Println("|'cpkgs init <dir-name> [-d]' - initialize a new project using cpkgs      |")
 			fmt.Println("|'cpkgs add <package-name>' - add C packages using cpkgs                  |")
 			fmt.Println("|'cpkgs install' - install all the packages in the current project        |")
 			fmt.Println("|'cpkgs run <file-name>' - run the file name using your selected compiler |")
 			fmt.Println("|'cpkgs help' - show this menu                                            |")
 			fmt.Println("---------------------------------------------------------------------------")
 			fmt.Print("\n")
+			break
 		}
 	default:
 		{
